sidebar improve, refactor misc
