+++ right/index.html
@@ -13,6 +13,6 @@
 
 	<body>
 	</body>
-
+<!-- 	<script id="gl-function"></script> -->
 	<script type="module" src="right.js"></script>
 </html>
